improve API for scenarios with different CLI
